+++ massiv/src/Data/Massiv/Array/Manifest/Boxed.hs
@@ -18,13 +18,20 @@
 --
 module Data.Massiv.Array.Manifest.Boxed
   ( B(..)
+  , BL(..)
   , N(..)
   , Array(..)
+  , wrapLazyArray
+  , unwrapLazyArray
   , unwrapNormalForm
   , evalNormalForm
   , unwrapArray
   , evalArray
+  , toLazyArray
+  , evalLazyArray
+  , forceLazyArray
   , unwrapMutableArray
+  , unwrapMutableLazyArray
   , evalMutableArray
   , unwrapNormalFormArray
   , evalNormalFormArray
@@ -32,13 +39,14 @@ module Data.Massiv.Array.Manifest.Boxed
   , evalNormalFormMutableArray
   , toBoxedVector
   , toBoxedMVector
+  , fromBoxedVector
+  , fromBoxedMVector
   , evalBoxedVector
   , evalBoxedMVector
   , evalNormalBoxedVector
   , evalNormalBoxedMVector
-  , unsafeBoxedArray
-  , unsafeNormalBoxedArray
-  , unsafeFromBoxedVector
+  , coerceBoxedArray
+  , coerceNormalBoxedArray
   , seqArray
   , deepseqArray
   ) where
@@ -80,6 +88,206 @@ sizeofArray (A.Array a#) = I# (sizeofArray# a#)
 sizeofMutableArray (A.MutableArray ma#) = I# (sizeofMutableArray# ma#)
 #endif
 
+----------------
+-- Boxed Lazy --
+----------------
+
+-- | Array representation for Boxed elements. This data structure is lazy with respect to
+-- its elements, but is strict with respect to the spine.
+data BL = BL deriving Show
+
+data instance Array BL ix e = BLArray { blComp   :: !Comp
+                                      , blSize   :: !(Sz ix)
+                                      , blOffset :: {-# UNPACK #-} !Int
+                                      , blData   :: {-# UNPACK #-} !(A.Array e)
+                                      }
+
+instance (Ragged L ix e, Show e) => Show (Array BL ix e) where
+  showsPrec = showsArrayPrec id
+  showList = showArrayList
+
+instance (Ragged L ix e, Show e) => Show (Array DL ix e) where
+  showsPrec = showsArrayPrec (computeAs BL)
+  showList = showArrayList
+
+instance Show e => Show (Array DS Ix1 e) where
+  showsPrec = showsArrayPrec (computeAs BL)
+  showList = showArrayList
+
+
+-- instance (Index ix, NFData e) => NFData (Array BL ix e) where
+--   rnf = (`deepseqArray` ())
+--   {-# INLINE rnf #-}
+
+instance (Index ix, Eq e) => Eq (Array BL ix e) where
+  (==) = eqArrays (==)
+  {-# INLINE (==) #-}
+
+instance (Index ix, Ord e) => Ord (Array BL ix e) where
+  compare = compareArrays compare
+  {-# INLINE compare #-}
+
+instance Index ix => Construct BL ix e where
+  setComp c arr = arr { blComp = c }
+  {-# INLINE setComp #-}
+
+  makeArrayLinear !comp !sz f = unsafePerformIO $ generateArrayLinear comp sz (pure . f)
+  {-# INLINE makeArrayLinear #-}
+
+instance Index ix => Source BL ix e where
+  unsafeLinearIndex (BLArray _ _sz o a) i =
+    INDEX_CHECK("(Source BL ix e).unsafeLinearIndex",
+                SafeSz . sizeofArray, A.indexArray) a (i + o)
+  {-# INLINE unsafeLinearIndex #-}
+
+  unsafeLinearSlice i k (BLArray c _ o a) = BLArray c k (o + i) a
+  {-# INLINE unsafeLinearSlice #-}
+
+
+instance Index ix => Resize BL ix where
+  unsafeResize !sz !arr = arr { blSize = sz }
+  {-# INLINE unsafeResize #-}
+
+instance Index ix => Extract BL ix e where
+  unsafeExtract !sIx !newSz !arr = unsafeExtract sIx newSz (toManifest arr)
+  {-# INLINE unsafeExtract #-}
+
+
+instance ( Index ix
+         , Index (Lower ix)
+         , Elt M ix e ~ Array M (Lower ix) e
+         , Elt BL ix e ~ Array M (Lower ix) e
+         ) =>
+         OuterSlice BL ix e where
+  unsafeOuterSlice arr = unsafeOuterSlice (toManifest arr)
+  {-# INLINE unsafeOuterSlice #-}
+
+instance ( Index ix
+         , Index (Lower ix)
+         , Elt M ix e ~ Array M (Lower ix) e
+         , Elt BL ix e ~ Array M (Lower ix) e
+         ) =>
+         InnerSlice BL ix e where
+  unsafeInnerSlice arr = unsafeInnerSlice (toManifest arr)
+  {-# INLINE unsafeInnerSlice #-}
+
+instance {-# OVERLAPPING #-} Slice BL Ix1 e where
+  unsafeSlice arr i _ _ = pure (unsafeLinearIndex arr i)
+  {-# INLINE unsafeSlice #-}
+
+
+instance Index ix => Manifest BL ix e where
+
+  unsafeLinearIndexM (BLArray _ _sz o a) i =
+    INDEX_CHECK("(Manifest BL ix e).unsafeLinearIndexM",
+                SafeSz . sizeofArray, A.indexArray) a (i + o)
+  {-# INLINE unsafeLinearIndexM #-}
+
+
+instance Index ix => Mutable BL ix e where
+  data MArray s BL ix e = MBLArray !(Sz ix) {-# UNPACK #-} !Int {-# UNPACK #-} !(A.MutableArray s e)
+
+  msize (MBLArray sz _ _) = sz
+  {-# INLINE msize #-}
+
+  unsafeThaw (BLArray _ sz o a) = MBLArray sz o <$> A.unsafeThawArray a
+  {-# INLINE unsafeThaw #-}
+
+  unsafeFreeze comp (MBLArray sz o ma) = BLArray comp sz o <$> A.unsafeFreezeArray ma
+  {-# INLINE unsafeFreeze #-}
+
+  unsafeNew sz = MBLArray sz 0 <$> A.newArray (totalElem sz) uninitialized
+  {-# INLINE unsafeNew #-}
+
+  initialize _ = return ()
+  {-# INLINE initialize #-}
+
+  unsafeLinearRead (MBLArray _ o ma) i =
+    INDEX_CHECK("(Mutable BL ix e).unsafeLinearRead",
+                SafeSz . sizeofMutableArray, A.readArray) ma (i + o)
+  {-# INLINE unsafeLinearRead #-}
+
+  unsafeLinearWrite (MBLArray _sz o ma) i e = e `seq`
+    INDEX_CHECK("(Mutable BL ix e).unsafeLinearWrite",
+                SafeSz . sizeofMutableArray, A.writeArray) ma (i + o) e
+  {-# INLINE unsafeLinearWrite #-}
+
+instance Index ix => Load BL ix e where
+  type R BL = M
+  size = blSize
+  {-# INLINE size #-}
+  getComp = blComp
+  {-# INLINE getComp #-}
+  loadArrayM !scheduler !arr = splitLinearlyWith_ scheduler (elemsCount arr) (unsafeLinearIndex arr)
+  {-# INLINE loadArrayM #-}
+
+instance Index ix => StrideLoad BL ix e
+
+instance Index ix => Stream BL ix e where
+  toStream = S.steps
+  {-# INLINE toStream #-}
+  toStreamIx = S.isteps
+  {-# INLINE toStreamIx #-}
+
+
+-- | Row-major sequential folding over a Boxed array.
+instance Index ix => Foldable (Array BL ix) where
+  fold = fold
+  {-# INLINE fold #-}
+  foldMap = foldMono
+  {-# INLINE foldMap #-}
+  foldl = lazyFoldlS
+  {-# INLINE foldl #-}
+  foldl' = foldlS
+  {-# INLINE foldl' #-}
+  foldr = foldrFB
+  {-# INLINE foldr #-}
+  foldr' = foldrS
+  {-# INLINE foldr' #-}
+  null (BLArray _ sz _ _) = totalElem sz == 0
+  {-# INLINE null #-}
+  length = totalElem . size
+  {-# INLINE length #-}
+  toList arr = build (\ c n -> foldrFB c n arr)
+  {-# INLINE toList #-}
+
+
+instance Index ix => Functor (Array BL ix) where
+  fmap f arr = makeArrayLinear (blComp arr) (blSize arr) (f . unsafeLinearIndex arr)
+  {-# INLINE fmap #-}
+
+instance Index ix => Traversable (Array BL ix) where
+  traverse = traverseA
+  {-# INLINE traverse #-}
+
+instance ( IsList (Array L ix e)
+         , Nested LN ix e
+         , Nested L ix e
+         , Ragged L ix e
+         ) =>
+         IsList (Array BL ix e) where
+  type Item (Array BL ix e) = Item (Array L ix e)
+  fromList = L.fromLists' Seq
+  {-# INLINE fromList #-}
+  toList = GHC.toList . toListArray
+  {-# INLINE toList #-}
+
+instance Num e => FoldNumeric BL e where
+  unsafeDotProduct = defaultUnsafeDotProduct
+  {-# INLINE unsafeDotProduct #-}
+  powerSumArray = defaultPowerSumArray
+  {-# INLINE powerSumArray #-}
+  foldArray = defaultFoldArray
+  {-# INLINE foldArray #-}
+
+instance Num e => Numeric BL e where
+  unsafeLiftArray = defaultUnsafeLiftArray
+  {-# INLINE unsafeLiftArray #-}
+  unsafeLiftArray2 = defaultUnsafeLiftArray2
+  {-# INLINE unsafeLiftArray2 #-}
+
+
+
 ------------------
 -- Boxed Strict --
 ------------------
@@ -88,28 +296,15 @@ sizeofMutableArray (A.MutableArray ma#) = I# (sizeofMutableArray# ma#)
 -- spine strict, but elements are strict to Weak Head Normal Form (WHNF) only.
 data B = B deriving Show
 
-data instance Array B ix e = BArray { bComp   :: !Comp
-                                    , bSize   :: !(Sz ix)
-                                    , bOffset :: {-# UNPACK #-} !Int
-                                    , bData   :: {-# UNPACK #-} !(A.Array e)
-                                    }
+newtype instance Array B ix e = BArray (Array BL ix e)
 
 instance (Ragged L ix e, Show e) => Show (Array B ix e) where
   showsPrec = showsArrayPrec id
   showList = showArrayList
 
-instance (Ragged L ix e, Show e) => Show (Array DL ix e) where
-  showsPrec = showsArrayPrec (computeAs B)
-  showList = showArrayList
-
-instance Show e => Show (Array DS Ix1 e) where
-  showsPrec = showsArrayPrec (computeAs B)
-  showList = showArrayList
-
-
-instance (Index ix, NFData e) => NFData (Array B ix e) where
-  rnf = (`deepseqArray` ())
-  {-# INLINE rnf #-}
+-- instance (Index ix, NFData e) => NFData (Array B ix e) where
+--   rnf = (`deepseqArray` ())
+--   {-# INLINE rnf #-}
 
 instance (Index ix, Eq e) => Eq (Array B ix e) where
   (==) = eqArrays (==)
@@ -120,24 +315,22 @@ instance (Index ix, Ord e) => Ord (Array B ix e) where
   {-# INLINE compare #-}
 
 instance Index ix => Construct B ix e where
-  setComp c arr = arr { bComp = c }
+  setComp c = coerce (\arr -> arr { blComp = c })
   {-# INLINE setComp #-}
 
-  makeArrayLinear !comp !sz f = unsafePerformIO $ generateArrayLinear comp sz (\ !i -> return $! f i)
+  makeArrayLinear !comp !sz f = unsafePerformIO $ generateArrayLinear comp sz (pure . f)
   {-# INLINE makeArrayLinear #-}
 
 instance Index ix => Source B ix e where
-  unsafeLinearIndex (BArray _ _sz o a) i =
-    INDEX_CHECK("(Source B ix e).unsafeLinearIndex",
-                SafeSz . sizeofArray, A.indexArray) a (i + o)
+  unsafeLinearIndex arr = unsafeLinearIndex (coerce arr :: Array BL ix e)
   {-# INLINE unsafeLinearIndex #-}
 
-  unsafeLinearSlice i k (BArray c _ o a) = BArray c k (o + i) a
+  unsafeLinearSlice i k arr = coerce (unsafeLinearSlice i k (coerce arr :: Array BL ix e))
   {-# INLINE unsafeLinearSlice #-}
 
 
 instance Index ix => Resize B ix where
-  unsafeResize !sz !arr = arr { bSize = sz }
+  unsafeResize sz = coerce (\arr -> arr { blSize = sz })
   {-# INLINE unsafeResize #-}
 
 instance Index ix => Extract B ix e where
@@ -170,47 +363,41 @@ instance {-# OVERLAPPING #-} Slice B Ix1 e where
 
 instance Index ix => Manifest B ix e where
 
-  unsafeLinearIndexM (BArray _ _sz o a) i =
-    INDEX_CHECK("(Manifest B ix e).unsafeLinearIndexM",
-                SafeSz . sizeofArray, A.indexArray) a (i + o)
+  unsafeLinearIndexM = coerce unsafeLinearIndexM
   {-# INLINE unsafeLinearIndexM #-}
 
 
 instance Index ix => Mutable B ix e where
-  data MArray s B ix e = MBArray !(Sz ix) {-# UNPACK #-} !Int {-# UNPACK #-} !(A.MutableArray s e)
+  newtype MArray s B ix e = MBArray (MArray s BL ix e)
 
-  msize (MBArray sz _ _) = sz
+  msize = msize . coerce
   {-# INLINE msize #-}
 
-  unsafeThaw (BArray _ sz o a) = MBArray sz o <$> A.unsafeThawArray a
+  unsafeThaw arr = MBArray <$> unsafeThaw (coerce arr)
   {-# INLINE unsafeThaw #-}
 
-  unsafeFreeze comp (MBArray sz o ma) = BArray comp sz o <$> A.unsafeFreezeArray ma
+  unsafeFreeze comp marr = BArray <$> unsafeFreeze comp (coerce marr)
   {-# INLINE unsafeFreeze #-}
 
-  unsafeNew sz = MBArray sz 0 <$> A.newArray (totalElem sz) uninitialized
+  unsafeNew sz = MBArray <$> unsafeNew sz
   {-# INLINE unsafeNew #-}
 
   initialize _ = return ()
   {-# INLINE initialize #-}
 
-  unsafeLinearRead (MBArray _ o ma) i =
-    INDEX_CHECK("(Mutable B ix e).unsafeLinearRead",
-                SafeSz . sizeofMutableArray, A.readArray) ma (i + o)
+  unsafeLinearRead ma = unsafeLinearRead (coerce ma)
   {-# INLINE unsafeLinearRead #-}
 
-  unsafeLinearWrite (MBArray _sz o ma) i e = e `seq`
-    INDEX_CHECK("(Mutable B ix e).unsafeLinearWrite",
-                SafeSz . sizeofMutableArray, A.writeArray) ma (i + o) e
+  unsafeLinearWrite ma i e = e `seq` unsafeLinearWrite (coerce ma) i e
   {-# INLINE unsafeLinearWrite #-}
 
 instance Index ix => Load B ix e where
   type R B = M
-  size = bSize
+  size = blSize . coerce
   {-# INLINE size #-}
-  getComp = bComp
+  getComp = blComp . coerce
   {-# INLINE getComp #-}
-  loadArrayM !scheduler !arr = splitLinearlyWith_ scheduler (elemsCount arr) (unsafeLinearIndex arr)
+  loadArrayM scheduler = coerce (loadArrayM scheduler)
   {-# INLINE loadArrayM #-}
 
 instance Index ix => StrideLoad B ix e
@@ -236,7 +423,7 @@ instance Index ix => Foldable (Array B ix) where
   {-# INLINE foldr #-}
   foldr' = foldrS
   {-# INLINE foldr' #-}
-  null (BArray _ sz _ _) = totalElem sz == 0
+  null arr = totalElem (size arr) == 0
   {-# INLINE null #-}
   length = totalElem . size
   {-# INLINE length #-}
@@ -245,7 +432,7 @@ instance Index ix => Foldable (Array B ix) where
 
 
 instance Index ix => Functor (Array B ix) where
-  fmap f arr = makeArrayLinear (bComp arr) (bSize arr) (f . unsafeLinearIndex arr)
+  fmap f arr = makeArrayLinear (getComp arr) (size arr) (f . unsafeLinearIndex arr)
   {-# INLINE fmap #-}
 
 instance Index ix => Traversable (Array B ix) where
@@ -287,15 +474,15 @@ instance Num e => Numeric B e where
 -- instance is required.
 data N = N deriving Show
 
-newtype instance Array N ix e = NArray { bArray :: Array B ix e }
+newtype instance Array N ix e = NArray { bArray :: Array BL ix e }
 
 instance (Ragged L ix e, Show e, NFData e) => Show (Array N ix e) where
   showsPrec = showsArrayPrec bArray
   showList = showArrayList
 
-instance (Index ix, NFData e) => NFData (Array N ix e) where
-  rnf (NArray barr) = barr `deepseqArray` ()
-  {-# INLINE rnf #-}
+-- instance (Index ix, NFData e) => NFData (Array N ix e) where
+--   rnf (NArray barr) = barr `deepseqArray` ()
+--   {-# INLINE rnf #-}
 
 instance (Index ix, NFData e, Eq e) => Eq (Array N ix e) where
   (==) = eqArrays (==)
@@ -307,17 +494,10 @@ instance (Index ix, NFData e, Ord e) => Ord (Array N ix e) where
 
 
 instance (Index ix, NFData e) => Construct N ix e where
-  setComp c (NArray arr) = NArray (arr {bComp = c})
+  setComp c (NArray arr) = NArray (arr {blComp = c})
   {-# INLINE setComp #-}
-  makeArray !comp !sz f =
-    unsafePerformIO $
-    generateArray
-      comp
-      sz
-      (\ !ix ->
-         let res = f ix
-          in res `deepseq` return res)
-  {-# INLINE makeArray #-}
+  makeArrayLinear !comp !sz f = unsafePerformIO $ generateArrayLinear comp sz (pure . f)
+  {-# INLINE makeArrayLinear #-}
 
 instance (Index ix, NFData e) => Source N ix e where
   unsafeLinearIndex (NArray arr) = unsafeLinearIndex arr
@@ -362,20 +542,20 @@ instance {-# OVERLAPPING #-} NFData e => Slice N Ix1 e where
 
 instance (Index ix, NFData e) => Manifest N ix e where
 
-  unsafeLinearIndexM (NArray arr) = unsafeLinearIndexM arr
+  unsafeLinearIndexM arr = unsafeLinearIndexM (coerce arr)
   {-# INLINE unsafeLinearIndexM #-}
 
 
 instance (Index ix, NFData e) => Mutable N ix e where
-  newtype MArray s N ix e = MNArray { bmArray :: MArray s B ix e }
+  newtype MArray s N ix e = MNArray (MArray s BL ix e)
 
-  msize = msize . bmArray
+  msize = msize . coerce
   {-# INLINE msize #-}
 
-  unsafeThaw (NArray arr) = MNArray <$> unsafeThaw arr
+  unsafeThaw arr = MNArray <$> unsafeThaw (coerce arr)
   {-# INLINE unsafeThaw #-}
 
-  unsafeFreeze comp (MNArray marr) = NArray <$> unsafeFreeze comp marr
+  unsafeFreeze comp marr = NArray <$> unsafeFreeze comp (coerce marr)
   {-# INLINE unsafeFreeze #-}
 
   unsafeNew sz = MNArray <$> unsafeNew sz
@@ -384,17 +564,17 @@ instance (Index ix, NFData e) => Mutable N ix e where
   initialize _ = return ()
   {-# INLINE initialize #-}
 
-  unsafeLinearRead (MNArray ma) = unsafeLinearRead ma
+  unsafeLinearRead ma = unsafeLinearRead (coerce ma)
   {-# INLINE unsafeLinearRead #-}
 
-  unsafeLinearWrite (MNArray ma) i e = e `deepseq` unsafeLinearWrite ma i e
+  unsafeLinearWrite ma i e = e `deepseq` unsafeLinearWrite (coerce ma) i e
   {-# INLINE unsafeLinearWrite #-}
 
 instance (Index ix, NFData e) => Load N ix e where
   type R N = M
-  size = bSize . bArray
+  size = blSize . coerce
   {-# INLINE size #-}
-  getComp = bComp . bArray
+  getComp = blComp . coerce
   {-# INLINE getComp #-}
   loadArrayM !scheduler !arr = splitLinearlyWith_ scheduler (elemsCount arr) (unsafeLinearIndex arr)
   {-# INLINE loadArrayM #-}
@@ -451,7 +631,7 @@ uninitialized = throw Uninitialized
 --
 -- @since 0.2.1
 unwrapArray :: Array B ix e -> A.Array e
-unwrapArray = bData
+unwrapArray = blData . coerce
 {-# INLINE unwrapArray #-}
 
 -- | /O(n)/ - Wrap a boxed array and evaluate all elements to a WHNF.
@@ -460,18 +640,64 @@ unwrapArray = bData
 evalArray ::
      Comp -- ^ Computation strategy
   -> A.Array e -- ^ Lazy boxed array from @primitive@ package.
-  -> Array B Ix1 e
-evalArray = fromArraySeq (\a -> a `seqArray` a)
+  -> Vector B e
+evalArray comp a = evalLazyArray $ setComp comp $ wrapLazyArray a
 {-# INLINE evalArray #-}
 
+
+-- | /O(1)/ - Unwrap boxed array. This will discard any possible slicing that has been
+-- applied to the array.
+--
+-- @since 0.6.0
+unwrapLazyArray :: Array BL ix e -> A.Array e
+unwrapLazyArray = blData
+{-# INLINE unwrapLazyArray #-}
+
+-- | /O(1)/ - Wrap a boxed array.
+--
+-- @since 0.6.0
+wrapLazyArray :: A.Array e -> Vector BL e
+wrapLazyArray a = BLArray Seq (SafeSz (sizeofArray a)) 0 a
+{-# INLINE wrapLazyArray #-}
+
+
+-- | /O(n)/ - Evaluate all elements of a boxed lazy array to weak head normal form
+--
+-- @since 0.6.0
+toLazyArray :: Array B ix e -> Array BL ix e
+toLazyArray = coerce
+{-# INLINE toLazyArray #-}
+
+-- | /O(n)/ - Evaluate all elements of a boxed lazy array to weak head normal form
+--
+-- @since 0.6.0
+evalLazyArray :: Index ix => Array BL ix e -> Array B ix e
+evalLazyArray arr = arr `seqArray` BArray arr
+{-# INLINE evalLazyArray #-}
+
+-- | /O(n)/ - Evaluate all elements of a boxed lazy array to normal form
+--
+-- @since 0.6.0
+forceLazyArray :: (NFData e, Index ix) => Array BL ix e -> Array N ix e
+forceLazyArray arr = arr `deepseqArray` NArray arr
+{-# INLINE forceLazyArray #-}
 
 -- | /O(1)/ - Unwrap mutable boxed array. This will discard any possible slicing that has been
 -- applied to the array.
 --
 -- @since 0.2.1
 unwrapMutableArray :: MArray s B ix e -> A.MutableArray s e
-unwrapMutableArray (MBArray _ _ marr) = marr
+unwrapMutableArray (MBArray (MBLArray _ _ marr)) = marr
 {-# INLINE unwrapMutableArray #-}
+
+
+-- | /O(1)/ - Unwrap mutable boxed lazy array. This will discard any possible slicing that has been
+-- applied to the array.
+--
+-- @since 0.6.0
+unwrapMutableLazyArray :: MArray s BL ix e -> A.MutableArray s e
+unwrapMutableLazyArray (MBLArray _ _ marr) = marr
+{-# INLINE unwrapMutableLazyArray #-}
 
 
 -- | /O(n)/ - Wrap mutable boxed array and evaluate all elements to WHNF.
@@ -481,7 +707,7 @@ evalMutableArray ::
      PrimMonad m
   => A.MutableArray (PrimState m) e -- ^ Mutable array that will get wrapped
   -> m (MArray (PrimState m) B Ix1 e)
-evalMutableArray = fromMutableArraySeq seq
+evalMutableArray = fmap MBArray . fromMutableArraySeq seq
 {-# INLINE evalMutableArray #-}
 
 -------------------
@@ -493,7 +719,7 @@ evalMutableArray = fromMutableArraySeq seq
 --
 -- @since 0.2.1
 unwrapNormalFormArray :: Array N ix e -> A.Array e
-unwrapNormalFormArray = bData . bArray
+unwrapNormalFormArray = blData . coerce
 {-# INLINE unwrapNormalFormArray #-}
 
 -- | /O(n)/ - Wrap a boxed array and evaluate all elements to a Normal Form (NF).
@@ -504,7 +730,7 @@ evalNormalFormArray ::
   => Comp -- ^ Computation strategy
   -> A.Array e -- ^ Lazy boxed array
   -> Array N Ix1 e
-evalNormalFormArray = fromArraySeq (\a -> a `deepseqArray` NArray a)
+evalNormalFormArray comp = forceLazyArray . setComp comp . wrapLazyArray
 {-# INLINE evalNormalFormArray #-}
 
 
@@ -513,7 +739,7 @@ evalNormalFormArray = fromArraySeq (\a -> a `deepseqArray` NArray a)
 --
 -- @since 0.2.1
 unwrapNormalFormMutableArray :: MArray s N ix e -> A.MutableArray s e
-unwrapNormalFormMutableArray (MNArray marr) = unwrapMutableArray marr
+unwrapNormalFormMutableArray = unwrapMutableLazyArray . coerce
 {-# INLINE unwrapNormalFormMutableArray #-}
 
 
@@ -528,36 +754,28 @@ evalNormalFormMutableArray marr = MNArray <$> fromMutableArraySeq deepseq marr
 {-# INLINE evalNormalFormMutableArray #-}
 
 
-----------------------
--- Helper functions --
-----------------------
+-- ----------------------
+-- -- Helper functions --
+-- ----------------------
 
 fromMutableArraySeq ::
      PrimMonad m
   => (e -> m () -> m a)
   -> A.MutableArray (PrimState m) e
-  -> m (MArray (PrimState m) B Ix1 e)
+  -> m (MArray (PrimState m) BL Ix1 e)
 fromMutableArraySeq with ma = do
   let !sz = sizeofMutableArray ma
   loopM_ 0 (< sz) (+ 1) (A.readArray ma >=> (`with` return ()))
-  return $! MBArray (SafeSz sz) 0 ma
+  return $! MBLArray (SafeSz sz) 0 ma
 {-# INLINE fromMutableArraySeq #-}
 
-fromArraySeq ::
-     (Array B Ix1 e -> a)
-  -> Comp
-  -> A.Array e
-  -> a
-fromArraySeq with comp barr = with (BArray comp (SafeSz (sizeofArray barr)) 0 barr)
-{-# INLINE fromArraySeq #-}
 
-
-seqArray :: Index ix => Array B ix a -> t -> t
+seqArray :: Index ix => Array BL ix a -> t -> t
 seqArray !arr t = foldlInternal (flip seq) () (flip seq) () arr `seq` t
 {-# INLINE seqArray #-}
 
 
-deepseqArray :: (NFData a, Index ix) => Array B ix a -> t -> t
+deepseqArray :: (NFData a, Index ix) => Array BL ix a -> t -> t
 deepseqArray !arr t = foldlInternal (flip deepseq) () (flip seq) () arr `seq` t
 {-# INLINE deepseqArray #-}
 
@@ -573,21 +791,21 @@ unwrapNormalForm = coerce
 --
 -- @since 0.5.0
 evalNormalForm :: (Index ix, NFData e) => Array B ix e -> Array N ix e
-evalNormalForm arr = arr `deepseqArray` NArray arr
+evalNormalForm (BArray arr) = arr `deepseqArray` NArray arr
 {-# INLINE evalNormalForm #-}
 
 -- | /O(1)/ - Converts a boxed `Array` into a `VB.Vector`.
 --
 -- @since 0.5.0
-toBoxedVector :: Index ix => Array B ix a -> VB.Vector a
+toBoxedVector :: Index ix => Array BL ix a -> VB.Vector a
 toBoxedVector arr = runST $ VB.unsafeFreeze . toBoxedMVector =<< unsafeThaw arr
 {-# INLINE toBoxedVector #-}
 
 -- | /O(1)/ - Converts a boxed `MArray` into a `VMB.MVector`.
 --
 -- @since 0.5.0
-toBoxedMVector :: Index ix => MArray s B ix a -> MVB.MVector s a
-toBoxedMVector (MBArray sz o marr) = MVB.MVector o (totalElem sz) marr
+toBoxedMVector :: Index ix => MArray s BL ix a -> MVB.MVector s a
+toBoxedMVector (MBLArray sz o marr) = MVB.MVector o (totalElem sz) marr
 {-# INLINE toBoxedMVector #-}
 
 -- | /O(n)/ - Convert a boxed vector and evaluate all elements to WHNF. Computation
@@ -595,9 +813,7 @@ toBoxedMVector (MBArray sz o marr) = MVB.MVector o (totalElem sz) marr
 --
 -- @since 0.5.0
 evalBoxedVector :: Comp -> VB.Vector a -> Array B Ix1 a
-evalBoxedVector comp v = arr `seqArray` arr
-  where
-    arr = setComp comp $ unsafeFromBoxedVector v
+evalBoxedVector comp = evalLazyArray . setComp comp . fromBoxedVector
 {-# INLINE evalBoxedVector #-}
 
 
@@ -606,40 +822,48 @@ evalBoxedVector comp v = arr `seqArray` arr
 --
 -- @since 0.5.0
 evalBoxedMVector :: PrimMonad m => MVB.MVector (PrimState m) a -> m (MArray (PrimState m) B Ix1 a)
-evalBoxedMVector (MVB.MVector o k ma) = do
-  let marr = MBArray (SafeSz k) o ma
-  loopM_ o (< k) (+ 1) (A.readArray ma >=> (`seq` pure ()))
-  pure marr
+evalBoxedMVector (MVB.MVector o k ma) =
+  let marr = MBArray (MBLArray (SafeSz k) o ma)
+   in marr <$ loopM_ o (< k) (+ 1) (A.readArray ma >=> (`seq` pure ()))
 {-# INLINE evalBoxedMVector #-}
 
 
--- | /O(n)/ - Cast a boxed vector without touching any elements. It is unsafe because it
--- violates the invariant that all elements of `B` array are in WHNF.
+-- | /O(1)/ - Cast a boxed vector without touching any elements.
 --
--- @since 0.5.0
-unsafeFromBoxedVector :: VB.Vector a -> Array B Ix1 a
-unsafeFromBoxedVector v =
+-- @since 0.6.0
+fromBoxedVector :: VB.Vector a -> Vector BL a
+fromBoxedVector v =
   runST $ do
     MVB.MVector o k ma <- VB.unsafeThaw v
-    unsafeFreeze Seq $ MBArray (SafeSz k) o ma
-{-# INLINE unsafeFromBoxedVector #-}
+    unsafeFreeze Seq $ MBLArray (SafeSz k) o ma
+{-# INLINE fromBoxedVector #-}
 
--- | /O(n)/ - Cast a boxed array. It is unsafe because it violates the invariant that all
--- elements of `N` array are in NF.
+
+-- | /O(1)/ - Convert mutable boxed vector to a lazy mutable boxed array. Both keep
+-- pointing to the same memory
 --
--- @since 0.5.0
-unsafeBoxedArray :: A.Array e -> Array B Ix1 e
-unsafeBoxedArray = fromArraySeq id Seq
-{-# INLINE unsafeBoxedArray #-}
+-- @since 0.6.0
+fromBoxedMVector :: MVB.MVector s a -> MArray s BL Ix1 a
+fromBoxedMVector (MVB.MVector o k ma) = MBLArray (SafeSz k) o ma
+{-# INLINE fromBoxedMVector #-}
 
 
--- | /O(n)/ - Cast a boxed array. It is unsafe because it violates the invariant that all
--- elements of `N` array are in NF.
+-- | /O(1)/ - Cast a boxed lazy array. It is unsafe because it can violate the invariant
+-- that all elements of `N` array are in NF.
 --
--- @since 0.5.0
-unsafeNormalBoxedArray :: Array B ix e -> Array N ix e
-unsafeNormalBoxedArray = coerce
-{-# INLINE unsafeNormalBoxedArray #-}
+-- @since 0.6.0
+coerceNormalBoxedArray :: Array BL ix e -> Array N ix e
+coerceNormalBoxedArray = coerce
+{-# INLINE coerceNormalBoxedArray #-}
+
+
+-- | /O(1)/ - Cast a boxed lazy array. It is unsafe because it can violate the invariant
+-- that all elements of `B` array are in WHNF.
+--
+-- @since 0.6.0
+coerceBoxedArray :: Array BL ix e -> Array B ix e
+coerceBoxedArray = coerce
+{-# INLINE coerceBoxedArray #-}
 
 -- | /O(n)/ - Convert mutable boxed vector and evaluate all elements to WHNF
 -- sequentially. Both keep pointing to the same memory
@@ -647,10 +871,9 @@ unsafeNormalBoxedArray = coerce
 -- @since 0.5.0
 evalNormalBoxedMVector ::
      (NFData a, PrimMonad m) => MVB.MVector (PrimState m) a -> m (MArray (PrimState m) N Ix1 a)
-evalNormalBoxedMVector (MVB.MVector o k ma) = do
-  let marr = MNArray (MBArray (SafeSz k) o ma)
-  loopM_ o (< k) (+ 1) (A.readArray ma >=> (`deepseq` pure ()))
-  pure marr
+evalNormalBoxedMVector (MVB.MVector o k ma) =
+  let marr = MNArray (MBLArray (SafeSz k) o ma)
+   in marr <$ loopM_ o (< k) (+ 1) (A.readArray ma >=> (`deepseq` pure ()))
 {-# INLINE evalNormalBoxedMVector #-}
 
 -- | /O(n)/ - Convert a boxed vector and evaluate all elements to WHNF. Computation
@@ -661,7 +884,6 @@ evalNormalBoxedVector :: NFData a => Comp -> VB.Vector a -> Array N Ix1 a
 evalNormalBoxedVector comp v =
   runST $ do
     MVB.MVector o k ma <- VB.unsafeThaw v
-    arr <- unsafeFreeze comp $ MBArray (SafeSz k) o ma
-    arr `deepseqArray` pure (NArray arr)
+    forceLazyArray <$> unsafeFreeze comp (MBLArray (SafeSz k) o ma)
 {-# INLINE evalNormalBoxedVector #-}